fixed problem with energy on repulsion start
git-svn-id: https://svn.vis.ethz.ch/svn/kob_gamelab09/trunk@1658 fc16ba98-2866-4055-be8f-df951a5456ab

+++ src/ProjectMagma/ProjectMagma/Simulation/Properties/Player/PlayerControllerProperty.cs
@@ -624,7 +624,7 @@ namespace ProjectMagma.Simulation
                             constants.GetFloat("island_repulsion_start_velocity_multiplier");
                         activeIsland.SetVector3("repulsion_velocity", currentVelocity + velocity);
 
-                        player.SetFloat("energy", player.GetInt("energy") - constants.GetInt("island_repulsion_start_energy_cost"));
+                        player.SetFloat("energy", player.GetFloat("energy") - constants.GetInt("island_repulsion_start_energy_cost"));
 
                         StartRepulsion();
                     }
